Final results: 300-sample Phase 2/3, 50-sample Phase 6, 5-method comparison
Phase 2 (300 samples): 87% at 0pm, 70% plateau at 10-70pm, 61% at 100pm, 30% at 200pm.
Phase 3 (300 samples): noised_ts 70%, midpoint 29%, reactant 6.3%, product 3%.
Phase 6 (50 samples): basin stable to 100pm (0 wrong TS), first diff TS at 200pm.
Phase 7 (5/7 methods): gad_small_dt (dt=0.005) best at 96%/10pm, 84%/50pm.
Adaptive dt hurts (clear negative result). Ping-pong dtype fix applied.

+++ src/gadplus/search/nr_gad_pingpong.py
@@ -190,7 +190,7 @@ def run_nr_gad_pingpong(
         if phase == "nr":
             # NR minimization: pure Newton descent on all modes
             n_nr_steps += 1
-            grad = -forces.reshape(-1)
+            grad = -forces.reshape(-1).to(evecs_vib_3N.dtype)
             delta_x, _info = nr_minimize_step(
                 grad, evals_vib, evecs_vib_3N,
                 max_step=cfg.nr_max_step,
